% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_scroller.R
\name{html_dependencies_scroller}
\alias{html_dependencies_scroller}
\title{scroller dependencies}
\usage{
html_dependencies_scroller(use_cdn = TRUE)
}
\arguments{
\item{use_cdn}{use CDN (by default) or use locally stored files.}
}
\description{
Get scroller html dependencies.
}
