% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_anchor.R
\name{use_scroller}
\alias{use_scroller}
\title{Use Scroller}
\usage{
use_scroller(
  animationLength = 750,
  easingFunction = "linear",
  scrollOffset = 0
)
}
\arguments{
\item{animationLength}{integrer. Lenght of the animation}

\item{easingFunction}{string. An effect from \href{http://gsgd.co.uk/sandbox/jquery/easing/}{jquery-easing}.}

\item{scrollOffset}{integrer. scrollOffset}
}
\description{
Use Scroller
}
\examples{

use_scroller()

}
